Moved to a mixin
