+++ app/scripts/views/details/details-history-view.js
@@ -6,8 +6,7 @@ const Locale = require('../../util/locale');
 const Alerts = require('../../comp/alerts');
 const FieldViewReadOnly = require('../fields/field-view-read-only');
 const FieldViewReadOnlyRaw = require('../fields/field-view-read-only-raw');
-const Tip = require('../../util/tip');
-const Timeouts = require('../../const/timeouts');
+const Copyable = require('../../mixins/copyable');
 
 const DetailsHistoryView = Backbone.View.extend({
     template: require('templates/details/details-history.hbs'),
@@ -212,39 +211,9 @@ const DetailsHistoryView = Backbone.View.extend({
                 this.closeHistory(true);
             }
         });
-    },
-
-    hideFieldCopyTip: function() {
-        if (this.fieldCopyTip) {
-            this.fieldCopyTip.hide();
-            this.fieldCopyTip = null;
-        }
-    },
-
-    fieldCopied: function(e) {
-        this.hideFieldCopyTip();
-        const fieldLabel = e.source.labelEl;
-        const clipboardTime = e.copyRes.seconds;
-        const msg = clipboardTime ? Locale.detFieldCopiedTime.replace('{}', clipboardTime)
-            : Locale.detFieldCopied;
-        let tip;
-        if (!this.isHidden()) {
-            tip = Tip.createTip(fieldLabel[0], {title: msg, placement: 'right', fast: true, force: true, noInit: true});
-            this.fieldCopyTip = tip;
-            tip.show();
-        }
-        setTimeout(() => {
-            if (tip) {
-                tip.hide();
-            }
-            this.fieldCopyTip = null;
-            if (e.source.model.name === '$Password' && AppSettingsModel.instance.get('lockOnCopy')) {
-                setTimeout(() => {
-                    Backbone.trigger('lock-workspace');
-                }, Timeouts.BeforeAutoLock);
-            }
-        }, Timeouts.CopyTip);
-    },
+    }
 });
+
+_.extend(DetailsHistoryView.prototype, Copyable);
 
 module.exports = DetailsHistoryView;
